requirement.txt
absl-py==1.4.0
altgraph==0.17.3
amqp==5.1.1
asgiref==3.5.0
astunparse==1.6.3
Babel==2.10.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==4.1.0
bs4==0.0.1
bytesbufio==1.0.3
cachetools==5.3.1
cascade==0.1
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.3
colorama==0.4.5
commonmark==0.9.1
comtypes==1.1.14
cssselect==1.2.0
cv==1.0.0
cycler==0.11.0
Cython==0.29.32
Django==4.0.3
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-qr-code==3.1.1
django-rest-framework==0.1.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docutils==0.18.1
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
flatbuffers==23.5.26
fonttools==4.31.2
future==0.18.2
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20
greenlet==2.0.1
grpcio==1.56.2
gTTS==2.3.0
h5py==3.9.0
idna==3.3
image==1.5.33
imaplib2==3.6
importlib-metadata==4.11.3
imutils==0.5.4
InvoiceGenerator==1.1.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.13.1
keyring==23.5.0
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.2
kombu==5.2.4
libclang==16.0.6
lxml==4.9.1
lyrics-extractor==3.0.1
Markdown==3.4.4
MarkupSafe==2.1.1
matplot==0.1.9
matplotlib==3.5.1
MouseInfo==0.1.3
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
patsy==0.5.2
pefile==2022.5.30
Pillow==9.1.0
pkginfo==1.8.2
plotly==5.9.0
plotly-express==0.4.1
plumbum==1.8.0
progressbar==2.5
prompt-toolkit==3.0.30
protobuf==4.23.4
pushbullet.py==0.12.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.2
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyloco==0.0.139
pymongo==4.0.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
python-pptx==0.6.21
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
qrplatba==0.5.0
readme-renderer==34.0
regex==2022.10.31
reportlab==3.6.10
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.1.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.1
segno==1.5.2
sgmllib3k==1.0.0
SimpleWebSocketServer==0.1.1
six==1.16.0
sklearn==0.0
sounddevice==0.4.5
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.44
sqlparse==0.4.2
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.1.0
tinysegmenter==0.3
tk==0.1.0
tkcalendar==1.6.1
tldextract==3.4.0
tqdm==4.64.1
twine==4.0.0
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
ushlex==0.99.1
vine==5.0.0
wavio==0.0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.2
winapps==0.2.0
wrapt==1.15.0
XlsxWriter==3.1.1
xlwt==1.3.0
zipp==3.8.0
